Move device profile classes to new package
